Fire STM32v2: Rename files for better conformance to naming conventions
