+++ nuttx/configs/fire-stm32v2/src/stm32_autoleds.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * configs/fire-stm32v2/src/up_autoleds.c
+ * configs/fire-stm32v2/src/stm32_autoleds.c
  *
  *   Copyright (C) 2012-2013, 2015 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
