+++ nuttx/configs/fire-stm32v2/src/stm32_nsh.c
@@ -1,10 +1,8 @@
 /****************************************************************************
- * configs/fire-stm32v2/src/up_composite.c
+ * config/fire-stm32v2/src/stm32_nsh.c
  *
  *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
- *
- * Configure and register the STM32 SPI-based MMC/SD block driver.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -41,18 +39,87 @@
 
 #include <nuttx/config.h>
 
+#include <stdbool.h>
 #include <stdio.h>
+#include <syslog.h>
+#include <errno.h>
 
+#include "stm32.h"
 #include "fire-internal.h"
 
 /****************************************************************************
- * Pre-Processor Definitions
+ * Pre-processor Definitions
  ****************************************************************************/
 /* Configuration ************************************************************/
-/* Device minor number */
 
-#ifndef CONFIG_SYSTEM_COMPOSITE_DEVMINOR1
-#  define CONFIG_SYSTEM_COMPOSITE_DEVMINOR1 0
+/* Assume that we support everything until convinced otherwise */
+
+#define HAVE_MMCSD    1
+#define HAVE_USBDEV   1
+#define HAVE_W25      1
+
+/* Configuration ************************************************************/
+/* SPI1 connects to the SD CARD (and to the SPI FLASH) */
+
+#define STM32_MMCSDSPIPORTNO   1  /* SPI1 */
+#define STM32_MMCSDSLOTNO      0  /* Only one slot */
+
+/* Can't support MMC/SD features if the SDIO peripheral is disabled */
+
+#ifndef CONFIG_STM32_SDIO
+#  undef HAVE_MMCSD
+#endif
+
+/* Can't support MMC/SD features if mountpoints are disabled */
+
+#ifdef CONFIG_DISABLE_MOUNTPOINT
+#  undef HAVE_MMCSD
+#endif
+
+/* Default MMC/SD minor number */
+
+#ifdef HAVE_MMCSD
+#  ifndef CONFIG_NSH_MMCSDMINOR
+#    define CONFIG_NSH_MMCSDMINOR 0
+#  endif
+
+/* Default MMC/SD SLOT number */
+
+#  if defined(CONFIG_NSH_MMCSDSLOTNO) && CONFIG_NSH_MMCSDSLOTNO != STM32_MMCSDSLOTNO
+#    error "Only one MMC/SD slot:  Slot 0"
+#    undef  CONFIG_NSH_MMCSDSLOTNO
+#    define CONFIG_NSH_MMCSDSLOTNO STM32_MMCSDSLOTNO
+#  endif
+
+#  ifndef CONFIG_NSH_MMCSDSLOTNO
+#    define CONFIG_NSH_MMCSDSLOTNO STM32_MMCSDSLOTNO
+#  endif
+#endif
+
+/* Can't support the W25 device if it SPI1 or W25 support is not enabled */
+
+#if !defined(CONFIG_STM32_SPI1) || !defined(CONFIG_MTD_W25)
+#  undef HAVE_W25
+#endif
+
+/* Can't support W25 features if mountpoints are disabled */
+
+#if defined(CONFIG_DISABLE_MOUNTPOINT)
+#  undef HAVE_W25
+#endif
+
+/* Default W25 minor number */
+
+#if defined(HAVE_W25) && !defined(CONFIG_NSH_W25MINOR)
+#  define CONFIG_NSH_W25MINOR 0
+#endif
+
+/* Can't support USB host or device features if the USB peripheral or the USB
+ * device infrastructure is not enabled
+ */
+
+#if !defined(CONFIG_STM32_USB) || !defined(CONFIG_USBDEV)
+#  undef HAVE_USBDEV
 #endif
 
 /****************************************************************************
@@ -60,26 +127,41 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: composite_archinitialize
+ * Name: nsh_archinitialize
  *
  * Description:
  *   Perform architecture specific initialization
  *
  ****************************************************************************/
 
-int composite_archinitialize(void)
+int nsh_archinitialize(void)
 {
-  /* If system/composite is built as an NSH command, then SD slot should
-   * already have been initized in nsh_archinitialize() (see up_nsh.c).  In
-   * this case, there is nothing further to be done here.
-   *
-   * NOTE: CONFIG_NSH_BUILTIN_APPS is not a fool-proof indication that NSH
-   * was built.
-   */
+#if defined(HAVE_MMCSD) || defined(HAVE_W25)
+  int ret;
+#endif
 
-#ifndef CONFIG_NSH_BUILTIN_APPS
-  return sd_mount(CONFIG_SYSTEM_COMPOSITE_DEVMINOR1);
-#else
+  /* Initialize and register the W25 FLASH file system. */
+
+#ifdef HAVE_W25
+  ret = stm32_w25initialize(CONFIG_NSH_W25MINOR);
+  if (ret < 0)
+    {
+      syslog(LOG_ERR, "ERROR: Failed to initialize W25 minor %d: %d\n",
+             CONFIG_NSH_W25MINOR, ret);
+      return ret;
+    }
+#endif
+
+  /* Initialize the SDIO-based MMC/SD slot */
+
+#ifdef HAVE_MMCSD
+  ret = stm32_sdinitialize(CONFIG_NSH_MMCSDMINOR);
+  if (ret < 0)
+    {
+      syslog(LOG_ERR, "ERROR: Failed to initialize MMC/SD slot %d: %d\n",
+             CONFIG_NSH_MMCSDSLOTNO, ret);
+      return ret;
+    }
+#endif
   return OK;
-#endif /* CONFIG_NSH_BUILTIN_APPS */
 }
