+++ nuttx/configs/fire-stm32v2/src/stm32_spi.c
@@ -1,6 +1,5 @@
 /************************************************************************************
- * configs/fire-stm32v2/src/up_spi.c
- * arch/arm/src/board/up_spi.c
+ * configs/fire-stm32v2/src/stm32_spi.c
  *
  *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -55,7 +54,7 @@
 #if defined(CONFIG_STM32_SPI1) || defined(CONFIG_STM32_SPI2)
 
 /************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ************************************************************************************/
 
 /* Enables debug output from this file (needs CONFIG_DEBUG too) */
