--- nuttx/configs/fire-stm32v2/src/up_buttons.c
@@ -1,157 +1,0 @@
-/****************************************************************************
- * configs/fire-stm32v2/src/board_buttons.c
- *
- *   Copyright (C) 2012, 2014-2015 Gregory Nutt. All rights reserved.
- *   Author: Gregory Nutt <gnutt@nuttx.org>
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
- * 3. Neither the name NuttX nor the names of its contributors may be
- *    used to endorse or promote products derived from this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
- * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *
- ****************************************************************************/
-
-/****************************************************************************
- * Included Files
- ****************************************************************************/
-
-#include <nuttx/config.h>
-
-#include <stdint.h>
-
-#include <nuttx/arch.h>
-#include <nuttx/board.h>
-#include <arch/board/board.h>
-
-#include "fire-internal.h"
-
-#ifdef CONFIG_ARCH_BUTTONS
-
-/****************************************************************************
- * Pre-processor Definitions
- ****************************************************************************/
-
-/****************************************************************************
- * Private Data
- ****************************************************************************/
-
-/****************************************************************************
- * Private Functions
- ****************************************************************************/
-
-/****************************************************************************
- * Public Functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: board_button_initialize
- *
- * Description:
- *   board_button_initialize() must be called to initialize button resources.
- *   After that, board_buttons() may be called to collect the current state of
- *   all buttons or board_button_irq() may be called to register button interrupt
- *   handlers.
- *
- ****************************************************************************/
-
-void board_button_initialize(void)
-{
-  /* Configure the GPIO pins as inputs.  NOTE that EXTI interrupts are
-   * configured for some pins but NOT used in this file
-   */
-
-  stm32_configgpio(GPIO_BTN_KEY1);
-  stm32_configgpio(GPIO_BTN_KEY2);
-}
-
-/****************************************************************************
- * Name: board_buttons
- ****************************************************************************/
-
-uint8_t board_buttons(void)
-{
-  uint8_t ret = 0;
-
-  /* Check that state of each key.  A LOW value means that the key is pressed, */
-
-  if (!stm32_gpioread(GPIO_BTN_KEY1))
-    {
-      ret |= BUTTON_KEY1_BIT;
-    }
-
-  if (!stm32_gpioread(GPIO_BTN_KEY2))
-    {
-      ret |= BUTTON_KEY2_BIT;
-    }
-
-  return ret;
-}
-
-/*****************************************************************************
- * Button support.
- *
- * Description:
- *   board_button_initialize() must be called to initialize button resources.
- *   After that, board_buttons() may be called to collect the current state of all
- *   buttons or board_button_irq() may be called to register button interrupt
- *   handlers.
- *
- *   After board_button_initialize() has been called, board_buttons() may be
- *   called to collect the state of all buttons.  board_buttons() returns an
- *   8-bit bit set with each bit associated with a button.  See the
- *   BUTTON_*_BIT and JOYSTICK_*_BIT definitions in board.h for the meaning
- *   of each bit.
- *
- *   board_button_irq() may be called to register an interrupt handler that will
- *   be called when a button is depressed or released.  The ID value is a
- *   button enumeration value that uniquely identifies a button resource. See
- *   the BUTTON_* and JOYSTICK_* definitions in board.h for the meaning of
- *   enumeration values.  The previous interrupt handler address is returned
- *   (so that it may restored, if so desired).
- *
- ****************************************************************************/
-
-#ifdef CONFIG_ARCH_IRQBUTTONS
-xcpt_t board_button_irq(int id, xcpt_t irqhandler)
-{
-  uint16_t gpio;
-
-  if (id == BUTTON_KEY1)
-    {
-      gpio = GPIO_KEY1;
-    }
-  else if (id == BUTTON_KEY2)
-    {
-      gpio = GPIO_KEY2;
-    }
-  else
-    {
-      return NULL;
-    }
-
-  return stm32_gpiosetevent(gpio, true, true, true, irqhandler);
-}
-#endif
-#endif /* CONFIG_ARCH_BUTTONS */
